+++ src/plugins/appInfo.js
@@ -1,0 +1,5 @@
+export default ({ app, Vue }) => {
+  Vue.prototype.pxsAppInfo = {
+    version: 'v0.0.5'
+  }
+}