+++ src/components/pxs-markdown.vue
@@ -1,13 +1,11 @@
 <template>
   <div>
-    <div v-if="text || !hideEmpty" class="round-borders q-pa-xs bg-green-1">
-      <div v-if="text">
-        <p v-for="(p, index) in text.split('\n')" :key="index"
-           :class="index > 0 ? 'q-mt-md' : ''"
-        >
-          {{p}}
-        </p>
-      </div>
+    <div v-if="text || !hideEmpty" class="round-borders q-pa-xs bg-light-blue-1">
+      <vue-markdown
+        v-if="text"
+        :source="text" table-class="markdownTable"
+        class="q-ma-xs"
+      />
       <div v-else style="color:gray;font-style:italic">
         {{emptyMessage}}
       </div>
@@ -16,6 +14,8 @@
 </template>
 
 <script>
+  import VueMarkdown from 'vue-markdown'
+
   export default {
     props: {
       text: {
@@ -32,6 +32,9 @@
         required: false,
         default: '(no description)'
       },
+    },
+    components: {
+      VueMarkdown,
     },
   }
 </script>